+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 Running changelog of releases since `2.0.1`
 
+## v3.0.0
+
+- Add support for `json+ion` error responses.
+- Add support to configure Authorization Headers. 
+- Rename `CreatedScoped` to `CreateScoped`.
+- Add `SendAsync` method in `BaseOktaClient`.
+
 ## v2.0.1